add preprocessing data functions

+++ src/data.py
@@ -52,7 +52,7 @@ def get_stock_data(cfg:dict):
     Returns:
         pd.DataFrame: Pandas dataframe of stock data sorted by their tickers 
     """
-    datapath = pathlib.Path(cfg["datapath"])
+    datapath = pathlib.Path(cfg["raw_datapath"])
 
     if exists(datapath):
         stock_data = pd.read_csv(datapath)
@@ -65,5 +65,50 @@ def get_stock_data(cfg:dict):
             stock['symbol'] = i
             stock_data = pd.concat([stock_data,stock])
         stock_data = stock_data.pivot(columns='symbol')
+        stock_data.to_csv(cfg['datapath'])
+    return stock_data
+
+def convert_multiindex(stock_data):
+    # check if all tickers are the same for all columns 
+    test_df = pd.DataFrame()
+    for i in ['^Date','^Open','^High','^Low','^Close','^Adj Close','^Volume']:
+        test_df = pd.concat([test_df,stock_data.filter(regex=i).iloc[0].reset_index(drop=True)],axis=1)
+    test_df['matching'] = test_df.eq(test_df.iloc[:, 0], axis=0).all(1)
+    assert not False in test_df['matching'], 'Tickers do not match in all columns'
+
+    # configure headers 
+    stock_data.drop(['Unnamed: 0'], axis=1,inplace=True)
+    header_cols = ['Date','Open','High','Low','Close','Adj Close','Volume']
+    length = len(stock_data.filter(like='Date').columns.tolist())
+    tickers = stock_data.filter(like='Date').iloc[0].tolist()
+    header_1 = []
+    for i in [length*[i] for i in header_cols]:
+        for j in i:
+            header_1.append(j)
+    header = [header_1, 
+            len(header_cols)*tickers]
+    stock_data.columns=header
+    stock_data = stock_data.iloc[1:].reset_index(drop=True) 
+    return stock_data 
+
+def drop_tickers(stock_data):
+    # 252 trading days in a year. If 10 years or more of missing data, drop the ticker
+    drop_ticker_list = []
+    for i in stock_data['Date'].columns.tolist():
+        if stock_data['Date'][i].isnull().sum() >= 2520:
+            drop_ticker_list.append(i)
+    stock_data = stock_data.drop(drop_ticker_list, axis=1, level = 1)
+    return stock_data
+
+def load_preprocessed_data(cfg:dict):
+    datapath = pathlib.Path(cfg["preprocessed_datapath"])
+
+    if exists(datapath):
+        stock_data = pd.read_csv(datapath)
+        stock_data = convert_multiindex(stock_data)
+    else:
+        stock_data = get_stock_data()
+        stock_data = convert_multiindex(stock_data)
+        stock_data = drop_tickers(stock_data)
         stock_data.to_csv(cfg['datapath'])
     return stock_data
